refactor: normalize component names and directories
